LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/dom/events/non-cancelable-when-passive/passive-wheel-event-listener-on-body-expected.txt


FAIL passive wheel event listener on body assert_equals: expected false but got true
